refactor(spending-modal): split SpendingModal into fields + hook (COS-58)
Pure refactor of the ~730-line SpendingModal.tsx into a readable shell that
composes sub-components — zero behaviour/render change.

- fields/: DateField, LabelField, AmountField, CategoryField, ReceiptField
- Toggle, useSpendingSubmit (amount eval + category resolution + dispatch)
- schema.ts (spendingSchema, SpendingForm, CategoryOption), helpers.ts
- shared state stays in the shell (register RHF, selectedCategory, combobox,
  receipt) and is passed by props, preserving persist-across-toggle behaviour

Cleanup on the touched files: drop manual useMemo (React Compiler), collapse
optional chains, receipt preview <img> -> next/image, remove dead eslint-disable.
MOCK notes now reference their de-mock tickets (COS-24/23/22).
