+++ front/src/components/spendings/common/spendingModal/helpers.ts
@@ -1,0 +1,20 @@
+export const FALLBACK_COLOR = "#94a3b8";
+
+export const humanSize = (bytes: number) => {
+  if (bytes < 1024) return `${bytes} o`;
+  if (bytes < 1_048_576) return `${Math.round(bytes / 1024)} Ko`;
+  return `${(bytes / 1_048_576).toFixed(1)} Mo`;
+};
+
+export const getRandomHexColor = () => {
+  const r = Math.floor(Math.random() * 255)
+    .toString(16)
+    .padStart(2, "0");
+  const g = Math.floor(Math.random() * 255)
+    .toString(16)
+    .padStart(2, "0");
+  const b = Math.floor(Math.random() * 255)
+    .toString(16)
+    .padStart(2, "0");
+  return `#${r}${g}${b}`;
+};
